# static/swagger.yaml
openapi: 3.0.0
info:
  title: API de Verificación de Rostros y Usuarios
  description: API para la verificación de rostros, inicio de sesión y registro de usuarios.
  version: 1.0.1
servers:
  - url: http://localhost:5000
    description: Servidor local
  - url: https://living-improved-ocelot.ngrok-free.app/
    description: Servidor en linea
...
paths:
...
  /api/signup/student-duplicate:
    post:
      summary: Verificar duplicado de usuario
      description: Verifica si un usuario ya existe en la base de datos basado en email, matnum o username.
      tags:
        - Registro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: Correo electrónico del usuario.
                matnum:
                  type: string
                  description: Número de matrícula del usuario.
                username:
                  type: string
                  description: Nombre de usuario.
              required:
                - email
                - matnum
                - username
      responses:
        '200':
          description: No se encontraron duplicados
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status_code:
                    type: integer
        '400':
          description: Solicitud incorrecta, campos faltantes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '409':
          description: Usuario ya registrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
                  duplicate_field:
                    type: string
                    description: Campo que causa la violación de unicidad (username, matnum o email).
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
...
  /api/login/teacher:
    post:
      summary: Iniciar sesión de profesor
      description: Permite a un profesor iniciar sesión proporcionando su número de trabajo y contraseña.
      tags:
          - Inicio de Sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                worknum:
                  type: string
                  description: Número de trabajo del profesor.
                password:
                  type: string
                  description: Contraseña del profesor.
              required:
                - worknum
                - password
      responses:
        '200':
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      face_img:
                        type: string
                        format: base64
                        description: Imagen de rostro en base64.
                  status_code:
                    type: integer
        '400':
          description: Solicitud incorrecta, campos faltantes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '402':
          description: Profesor no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '403':
          description: Contraseña incorrecta
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
  /api/face/verify:
    post:
      summary: Verificar coincidencia de rostros
      description: Compara dos imágenes de rostros y determina si coinciden.
      tags:
        - Rostros
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cap_frame:
                  type: string
                  format: base64
                  description: Imagen capturada en formato base64.
                ref_frame:
                  type: string
                  format: base64
                  description: Imagen de referencia en formato base64.
              required:
                - cap_frame
                - ref_frame
      responses:
        '200':
          description: Coincidencia de rostros exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      match:
                        type: boolean
                  status_code:
                    type: integer
        '400':
          description: Solicitud incorrecta, falta alguna imagen o hay un error en los datos proporcionados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
...
  /api/exam/results:
    get:
      summary: Retrieve exam results
      description: Retrieves all results for an exam based on exam_id.
      tags:
        - Exámenes
      parameters:
        - in: query
          name: exam_id
          required: true
          schema:
            type: string
          description: ID of the exam.
      responses:
        '200':
          description: Exam results retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  status_code:
                    type: integer
        '400':
          description: Missing exam_id parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
  /api/exam/modify-result:
    post:
      summary: Modify exam results
      description: Modifies multiple exam results with the provided details.
      tags:
        - Exámenes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                results:
                  type: array
                  items:
                    type: object
                    properties:
                      result_id:
                        type: integer
                        description: ID of the exam result (optional for new results).
                      score:
                        type: number
                        format: float
                        description: Score of the exam. Must be a number between 0 and 100.
                      student_id:
                        type: integer
                        description: ID of the student (required for new results).
                      exam_id:
                        type: integer
                        description: ID of the exam (required for new results).
                      class_id:
                        type: integer
                        description: ID of the class (required for new results).
                  required:
                    - score
      responses:
        '200':
          description: All exam results processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                  status_code:
                    type: integer
        '400':
          description: Missing required fields or invalid score value. Score must be a number between 0 and 100.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
...